Se crean todas las apis, quedan pendientes algunas correciones de seguridad, pero ya son funcionales

+++ backend/main/login/models.py
@@ -21,7 +21,7 @@ class Trabajador(models.Model):
     cedula = models.IntegerField(verbose_name="cedula")
     direccion = models.TextField(verbose_name="dirección")
     correro = models.EmailField(max_length=255, verbose_name="correo")
-    telefono = models.IntegerField(verbose_name="telefono")
+    telefono = models.BigIntegerField(verbose_name="telefono")
     cargo = models.CharField(max_length=255, verbose_name="cargo")
 
     def __str__(self):
